A bunch more layer reorg, splitting many layers into own files. Improve torchscript compatibility.
