+++ timm/models/layers/drop.py
